Remove erroneous URL params & hashes from img src
Fix URL params & hashes making it through when dealing wtih image
caching and optimization.

Add temporary workaround for dealing with doubly-encoded ampersands,
particularly in this instance of an <img> src.

URL params & hashes shouldn't be making it through to the caching &
optimization logic -- also due to ampersand encoding bug, the 'static'
URL param would only work if it was the only param.